# Conflicts:
#	NFComm/NFCore/NFIObject.h
#	NFComm/NFPluginModule/NFIPluginManager.h

+++ install.sh
@@ -1,7 +1,16 @@
+# make sure cmake is installed
+cmake --version
+if [ $? -ne 0 ]; then
+    echo "[ERROR] Please install cmake first."
+    echo "[ubuntu] apt-get install cmake or [centos] yum install cmake."
+    exit 1
+fi
+
+
 rm -rf ./cmake.dir/tmp
 mkdir ./cmake.dir/tmp
 cp -R -f ./cmake.dir/develop/* ./cmake.dir/tmp/
-find ./cmake.dir/tmp -name "*.text" -exec rename -f 's/.text/.txt/' {} \;
+python -mplatform | grep -qi Ubuntu && find ./cmake.dir/tmp -name "*.text" -exec rename -f 's/.text/.txt/' {} \; || find ./cmake.dir/tmp -name "*.text" -exec rename .text .txt {} \;
 cp -R -f ./cmake.dir/tmp/* ./
 
 # compile dep libraries
